Renamed dualize to duality.
Also moved a deprecated transformation into duality, which
creates a Lagrangian dual.
