+++ pyomo/bilevel/plugins/dual.py
@@ -61,7 +61,7 @@ class LinearDual_BilevelTransformation(Base_BilevelTransformation):
         """
         Generate the dual of a submodel
         """ 
-        transform = TransformationFactory('core.linear_dual')
+        transform = TransformationFactory('duality.linear_dual')
         return transform._dualize(submodel, unfixed)
 
     def _xfrm_bilinearities(self, dual):
