+++ pyomo/duality/collect.py
