+++ pyomo/duality/tests/t5_linear_dual.lp
